requirements.txt
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
courlan==1.3.2
cryptography==46.0.4
dateparser==1.3.0
distro==1.9.0
dotenv==0.9.9
fire==0.7.1
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
numpy==2.4.2
openai==2.17.0
opencv-python-headless==4.13.0.92
pandoc==2.4
pdf2docx==0.5.8
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
playwright==1.58.0
plumbum==1.10.0
ply==3.11
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyMuPDF==1.26.7
pypandoc==1.16.2
pyparsing==3.3.2
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
regex==2026.1.15
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
termcolor==3.3.0
tld==0.13.1
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
